beatport/requirements.txt
requests==2.31.0
fastapi==0.100.1
uvicorn==0.23.2
pydantic-settings==2.0.2